TYP: minor fixes related to ``errstate`` (#29914)
